requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
yt-dlp==2024.3.10
requests==2.31.0
PyYAML==6.0.1
instaloader==4.10.2
python-dotenv 